+++ flight/PiOS/STM32F30x/pios_dma_hooks.c
@@ -26,6 +26,7 @@
  * with this program; if not, write to the Free Software Foundation, Inc., 
  * 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  */
+
 #if defined(PIOS_INCLUDE_DMA_CB_SUBSCRIBING_FUNCTION)
 #include <pios_dma.h>
 
@@ -59,13 +60,14 @@ void DMA2_Channel5_IRQHandler() __attribute__ ((alias ("PIOS_DMA_25_irq_handler"
  * @brief calls the handlers associated with a given index (from PIOS_DMA_CHANNELS)
  * @param[in] DMA channel index (PIOS_DMA_CHANNELS)
  */
-static inline PIOS_DMA_Mapper(uint8_t index) {
-	if(pios_dma_handler_map[index][0]==NULL)
+static inline PIOS_DMA_Mapper(uint32_t index)
+{
+	if (pios_dma_handler_map[index][0] == NULL )
 		PIOS_DMA_Default_Handler();
-	for(uint8_t i=0;i<PIOS_DMA_MAX_HANDLERS_PER_CHANNEL;++i){
-		if(pios_dma_handler_map[index][i]==NULL)
+	for (uint32_t i = 0; i < PIOS_DMA_MAX_HANDLERS_PER_CHANNEL; ++i) {
+		if (pios_dma_handler_map[index][i] == NULL )
 			return;
-		else{
+		else {
 			pios_dma_handler_map[index][i]();
 		}
 	}
